Order Trouble Detection Improvement
Small change to the way the script detects order troubles. Instead of basing it on the notes, the order trouble bases it on the order 'type' from the JSON. That way, customer service does not have to send notes for it to be considered an order trouble - only send as a fault.