batch create: add blank pdf for odd panels that have no match
if an odd pdf added to a batch doesn't have a match at the next file, the script will add a blank pdf from the assets folder. The blank pdf will help keep the structure of the batch so things can be organized by due date. Otherwise, caldera may reorder full orders to save space and due dates may shift.

The blank pdf is copied from the generic version based on the height. The height is pulled from the LAST item in the batch list, before the current item is added. So if iterating through the available pdfs, the script looks at the sorted list. If an item is odd paneled, it will look for an odd panel next that matches in height. If the next item is NOT a matching odd panel (and won't make the batch too long),  then the blank pdf is added.