batchCreate: fixed adding blank PDF
In order to respect priorities, we add in a blank PDF so caldera doesn't reorganize pdfs to other priorities.

Adding blank PDFs was being skipped if the last item to be added in a section was odd. Fixed.
Adding blank PDFs during the middle of batches was assigning the wrong order number to the blank PDF. Fixed.