wallpaperSorterFunctions: duplicate samples with qty>1
Historically, if a sample has a qty greater than 1, caldera will manage making multiple copies of it. However, that causes the scripts batch length tracking to be off since a sample's length will read as only 9.5, but may actually be 9.5*qty. 

Now, upon unzipping, renaming, and sorting PDFs, the script will also split samples that are qty>1 into multiple samples with a qty of 1 each. To keep each filename unique, the template name is appended with `Template Name(counter)`.

This is also done in preparation for sorting samples for the cutter so they are easier to stack for the employees cutting them down.