wallpaperSorterFunctions: improved wallpaper sorting
Sorting is now gracefully handled and far more fault tolerant.

Sorting will work one package at a time, so in the event that an error is occurred, the user won't be stuck with an entire directory of packages that are partially unzipped, partially named, and a huge mess.

Sorter unzips a package, renames the contents of that package, sorts the PDFs, then moves to the next one. Meanwhile, the original zip file is left untouched.

If the sorter completes the entire run without error, then the XMLs will be removed, followed by the JSONs, followed by empty folders, followed by zipped packages.